requirements.txt
streamlit
langchain
pypdf
chromadb
openai
python-dotenv
sentence-transformers
google-generativeai